requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

dateparser == 1.1.1
emoji_country_flag == 1.3.1
python_dotenv == 0.20.0
python_telegram_bot == 20.0a2
requests == 2.27.1
